# Benchmark inplace or view fallback templated or boxed