--- a/tools/autograd/templates/ADInplaceOrViewType.cpp
+++ b/tools/autograd/templates/ADInplaceOrViewType.cpp
@@ -1,6 +1,7 @@
 #include "torch/csrc/autograd/VariableTypeUtils.h"
 
 #include <torch/library.h>
+#include <torch/csrc/autograd/autograd_not_implemented_fallback.h>
 
 
 #include <ATen/RedispatchFunctions.h>