--- a/torch/csrc/autograd/autograd_not_implemented_fallback.cpp
+++ b/torch/csrc/autograd/autograd_not_implemented_fallback.cpp
@@ -138,7 +138,7 @@ void autogradNotImplementedFallbackImpl(const c10::OperatorHandle& op, c10::Disp
   } else {
     // If neither in-place nor view
     at::AutoDispatchBelowADInplaceOrView guard;
-    op.redispatchBoxed(dispatch_keys & c10::after_autograd_keyset, stack);
+    op.redispatchBoxed(dispatch_keys & c10::after_ADInplaceOrView_keyset, stack);
   }
   #ifndef NDEBUG
   _foreach_tensor([&](size_t idx_tensor, size_t _, const at::Tensor& t) {
@@ -184,6 +184,130 @@ void autogradNotImplementedFallbackImpl(const c10::OperatorHandle& op, c10::Disp
 
 torch::CppFunction autogradNotImplementedFallback() {
   return torch::CppFunction::makeFromBoxedFunction<&autogradNotImplementedFallbackImpl>();
+}
+
+// Unfortunately we have to do this lookup...
+static const std::vector<std::string> NEEDS_METADATA_CHANGE = {"aten::view_as_complex", "aten::view_as_real", "aten::_conj", "aten::_neg_view"};
+
+// [WIP] What about in-place views
+void ADInplaceOrViewFallbackImpl(const c10::OperatorHandle& op, c10::DispatchKeySet dispatch_keys, torch::jit::Stack* stack) {
+  const auto& schema = op.schema();
+  const auto& op_name = schema.operator_name().name;
+  const auto& arguments = schema.arguments();
+  const auto& returns = schema.returns();
+  const auto num_arguments = arguments.size();
+  const auto num_returns = returns.size();
+  const auto stack_start = stack->size() - num_arguments;
+  bool any_is_inplace = false;
+
+  std::vector<size_t> tensors_to_increment_version_on_stack;
+  at::Tensor aliased_input;
+
+  int64_t aliased_output_idx = -1;
+  for (const auto i : c10::irange(num_returns)) {
+    const auto& alias_info = returns[i].alias_info();
+    if (alias_info.has_value() && !alias_info->isWrite()) {
+      // TODO: use torch check instead of AT_ASSERT
+      AT_ASSERT(aliased_output_idx == -1, "Expected a single output to be aliased, but observed at least 2 outputs with alias info");
+      aliased_output_idx = i;
+    }
+  }
+
+  int64_t aliased_input_idx = -1;
+  for (const auto i : c10::irange(num_arguments)) {
+    const auto& alias_info = arguments[i].alias_info();
+    if (alias_info.has_value()) {
+      if (!alias_info->isWrite()) {
+        aliased_input_idx = i;
+        // TODO: use torch check instead of AT_ASSERT
+        AT_ASSERT(!aliased_input.defined(), "Expected a single input to be aliased, but observed at least 2 input with alias info");
+        const c10::IValue& aliased_input_iv = (*stack)[stack_start + i]; // get a reference to an ivalue on the stack
+        // TODO: use torch check instead of AT_ASSERT
+        AT_ASSERT(aliased_input_iv.isTensor());
+        // copy assignment
+        aliased_input = aliased_input_iv.toTensor();
+      } else {
+        tensors_to_increment_version_on_stack.push_back(i);
+        any_is_inplace = true;
+      }
+    }
+  }
+  TORCH_INTERNAL_ASSERT((aliased_input_idx == -1 && aliased_output_idx == -1) ||
+    (aliased_input_idx != -1 && aliased_output_idx != -1))
+  const bool is_view = aliased_input_idx != -1;
+  const bool need_view_func = is_view
+                         && (std::find(NEEDS_METADATA_CHANGE.begin(), NEEDS_METADATA_CHANGE.end(), op_name) != NEEDS_METADATA_CHANGE.end()
+                             || !aliased_input.unsafeGetTensorImpl()->support_as_strided());
+  std::function<at::Tensor(const at::Tensor&)> view_func = nullptr;
+  std::vector<c10::IValue> stack_args_copy;
+  if (any_is_inplace || need_view_func) {
+    stack_args_copy = std::vector<c10::IValue>(stack->begin() + stack_start, stack->end());
+  }
+
+  if (need_view_func) {
+    // TODO: simplifications here, if we can assume the view is always the first return, and that
+    // there will only be a single return
+    view_func = [=](const at::Tensor& t) {
+      // We have to make another copy because the same lambda can be used twice! e.g., double backward
+      std::vector<c10::IValue> stack_args_copy_copy(stack_args_copy);
+      stack_args_copy_copy.at(stack_args_copy_copy.size() - num_arguments + aliased_input_idx) = t;
+      op.callBoxed(&stack_args_copy_copy);
+      return stack_args_copy_copy[stack_args_copy_copy.size() - num_returns + aliased_output_idx].toTensor();
+    };
+  }
+
+  {
+    at::AutoDispatchBelowADInplaceOrView guard;
+    op.redispatchBoxed(dispatch_keys & c10::after_ADInplaceOrView_keyset, stack);
+  }
+
+  if (any_is_inplace) {
+    for (const auto i : c10::irange(tensors_to_increment_version_on_stack.size())) {
+      increment_version(stack_args_copy[tensors_to_increment_version_on_stack[i]].toTensor());
+    }
+  }
+
+  if (is_view) {
+    const c10::IValue& aliased_output_iv = (*stack)[stack->size() - num_returns + aliased_output_idx];
+    if (aliased_output_iv.isTensorList()) {
+      auto aliased_output = aliased_output_iv.toTensorVector();
+      // Only allow rebasing of the history if we return a single Tensor
+      // If view was created in a no grad or inference mode block, raise an error
+      // See NOTE [ View + Inplace detection ] for more details about this logic
+      auto result = as_view(
+        /* base=*/aliased_input,
+        /* output=*/aliased_output,
+        /* is_bw_differentiable=*/true,
+        /* is_fw_differentiable=*/true,
+        /* creation_meta=*/InferenceMode::is_enabled() ? CreationMeta::INFERENCE_MODE : (at::GradMode::is_enabled() ? CreationMeta::MULTI_OUTPUT_NODE : CreationMeta::NO_GRAD_MODE));
+        // ^ pass in creation meta unecessarily even if not isDifferentiableType
+      stack->at(stack->size() - num_returns + aliased_output_idx) = result;
+    } else {
+      AT_ASSERT(aliased_output_iv.isTensor());
+      const at::Tensor& aliased_output = aliased_output_iv.toTensor();
+      if (isDifferentiableType(aliased_output.scalar_type())) {
+        auto result = as_view(
+          /* base=*/aliased_input,
+          /* output=*/aliased_output,
+          /* is_bw_differentiable=*/true,
+          /* is_fw_differentiable=*/true,
+          /* view_func=*/view_func,
+          /* creation_meta=*/InferenceMode::is_enabled() ? CreationMeta::INFERENCE_MODE : (at::GradMode::is_enabled() ? CreationMeta::DEFAULT : CreationMeta::NO_GRAD_MODE));
+        stack->at(stack->size() - num_returns + aliased_output_idx) = result;
+      } else {
+        auto result = as_view(
+        /* base=*/aliased_input,
+        /* output=*/aliased_output,
+        /* is_bw_differentiable=*/false,
+        /* is_fw_differentiable=*/false);
+        stack->at(stack->size() - num_returns + aliased_output_idx) = result;
+      }
+    }
+  }
+}
+
+torch::CppFunction ADInplaceOrViewFallback() {
+  return torch::CppFunction::makeFromBoxedFunction<&ADInplaceOrViewFallbackImpl>();
 }
 
 }} // namespace torch::autograd