This PR implements both templated and boxed version:

Benchmarks comparing templated and boxed versions in its current state.
We test for baseline (addition out-of-place), view op (just alias), in-place (addition in-place)
In the measurements below always register boxed Autograd kernel, but for...
 - boxed => boxed kernel for ADInplaceOrView
 - templated => templated kernel for ADInplaceOrView
 - codegen => copy-pasted codegened kernel for ADInplaceOrview
 - fallthrough => don't register ADInplaceOrView
 - we expect templated to be faster than boxed in all cases because
   - we can precompute values on construction
   - no need to incur reference count bumps from copying the stack in the in-place case
```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
baseline                                 731 ns          731 ns       942423
boxed                                   1561 ns         1561 ns       448876
templated                               1568 ns         1568 ns       447501
baseline_view                           18.3 ns         18.3 ns     37993160
boxed_view                              1031 ns         1031 ns       687561
templated_view                           994 ns          994 ns       705317
codegen_view                             983 ns          983 ns       704384
fallthrough_view                         595 ns          595 ns      1178183
baseline_inplace                         352 ns          352 ns      1945817
boxed_inplace                           1339 ns         1339 ns       525608
templated_inplace                       1247 ns         1247 ns       548316
codegen_inplace                         1217 ns         1217 ns       548446
fallthrough_inplace                     1499 ns         1499 ns       527335 <---- if fallthrough_inplace is run directly after codegen_inplace, it is slow
fallthrough_inplace_a_second_time       1143 ns         1142 ns       627750 <---- same test again is faster!
```
 - As shown above, there is some weird interactions that depend on the orderings of the timings, but running the test a second time may correct the measurement...
<details>
<summary>
Code
</summary>

```c++
#include <torch/extension.h>
#include <torch/csrc/autograd/autograd_not_implemented_fallback.h>
#include <benchmark/benchmark.h>

using namespace torch;
using namespace aten;
using namespace torch::autograd;

torch::Tensor my_test_op(const torch::Tensor& self, const torch::Tensor& other) {
  return self + other;
}

torch::Tensor my_test_view_op2(const torch::Tensor& self) {
  return self;
}

const torch::Tensor& my_test_inplace_op(const torch::Tensor& self, const torch::Tensor& other) {
  return self.add_(other);
}

// We want to isolate the time it takes to go through Autograd kernel for in-place op and
// fallthrough the InplaceOrView kernel
const torch::Tensor&  my_test_inplace_op_for_fallthrough(const torch::Tensor& self, const torch::Tensor& other ) {
  return self.add_(other);
}

// Registering this inplace op fallthrough shouldn't really impact perf
// const torch::Tensor&  my_test_inplace_op_fallthrough(const torch::Tensor& self, const torch::Tensor& other ) {
//   at::AutoDispatchBelowADInplaceOrView guard;
//   return my_test_inplace_op_for_fallthrough(self, other);
// }

// torch/include/ATen/RedispatchFunctions.h
// aten::alias(Tensor(a) self) -> Tensor(a)
// at::Tensor alias_redispatch(c10::DispatchKeySet dispatchKeySet, const at::Tensor & self) {
//     return at::_ops::alias::redispatch(dispatchKeySet, self);
// }

// Codegened ADInplaceOrView kernel
at::Tensor aliasInplaceOrViewCodegened(c10::DispatchKeySet ks, const at::Tensor & self) {
  auto _tmp = ([&]() {
    at::AutoDispatchBelowADInplaceOrView guard;
    return my_test_view_op2(self); // Lets just not do any redispatch
    // return alias_redispatch(ks & c10::after_ADInplaceOrView_keyset, self);
  })();
  std::function<at::Tensor(const at::Tensor&)> func=nullptr;
  if (false || !self.unsafeGetTensorImpl()->support_as_strided()) {
    func = [=](const at::Tensor& input_base) {
      return at::alias(input_base);
    };
  }
  auto result = as_view(/* base */ self, /* output */ _tmp, /* is_bw_differentiable */ true, /* is_fw_differentiable */ true, /* view_func */ func, /* creation_meta */ InferenceMode::is_enabled() ? CreationMeta::INFERENCE_MODE : (at::GradMode::is_enabled() ? CreationMeta::DEFAULT : CreationMeta::NO_GRAD_MODE));
  return result;
}

const at::Tensor & addInplaceCodegened(c10::DispatchKeySet ks, const at::Tensor & self, const at::Tensor & other) {
  {
    at::AutoDispatchBelowADInplaceOrView guard;
    my_test_inplace_op(self, other);
  }
  increment_version(self);
  return self;
}

Tensor dispatch_to_baseline(const Tensor& self, const Tensor& other) {
  at::AutoDispatchBelowADInplaceOrView guard;
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_op_baseline", "")
    .typed<decltype(my_test_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_boxed(const Tensor& self, const Tensor& other) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_op_boxed", "")
    .typed<decltype(my_test_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_templated(const Tensor& self, const Tensor& other) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_op_templated", "")
    .typed<decltype(my_test_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_baseline_view(const Tensor& self) {
  at::AutoDispatchBelowADInplaceOrView guard;
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_view_op2_baseline", "")
    .typed<decltype(my_test_view_op2)>();
  return op.call(self);
}

Tensor dispatch_to_boxed_view(const Tensor& self) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_view_op2_boxed", "")
    .typed<decltype(my_test_view_op2)>();
  return op.call(self);
}

Tensor dispatch_to_templated_view(const Tensor& self) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_view_op2_templated", "")
    .typed<decltype(my_test_view_op2)>();
  return op.call(self);
}

Tensor dispatch_to_codegen_view(const Tensor& self) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_view_op2_codegen", "")
    .typed<decltype(my_test_view_op2)>();
  return op.call(self);
}

Tensor dispatch_to_fallthrough_view(const Tensor& self) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_view_op2_fallthrough", "")
    .typed<decltype(my_test_view_op2)>();
  return op.call(self);
}

Tensor dispatch_to_baseline_inplace(const Tensor& self, const Tensor& other) {
  at::AutoDispatchBelowADInplaceOrView guard;
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_inplace_op_baseline", "")
    .typed<decltype(my_test_inplace_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_boxed_inplace(const Tensor& self, const Tensor& other) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_inplace_op_boxed", "")
    .typed<decltype(my_test_inplace_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_templated_inplace(const Tensor& self, const Tensor& other) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_inplace_op_templated", "")
    .typed<decltype(my_test_inplace_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_codegen_inplace(const Tensor& self, const Tensor& other) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_inplace_op_codegen", "")
    .typed<decltype(my_test_inplace_op)>();
  return op.call(self, other);
}

Tensor dispatch_to_fallthrough_inplace(const Tensor& self, const Tensor& other) {
  static auto op = torch::Dispatcher::singleton()
    .findSchemaOrThrow("my_ops::my_test_inplace_op_fallthrough", "")
    .typed<decltype(my_test_inplace_op_for_fallthrough)>();
  return op.call(self, other);
}

TORCH_LIBRARY(my_ops, m) {
  m.def("my_test_op_baseline(Tensor self, Tensor other) -> Tensor ");
  m.def("my_test_op_boxed(Tensor self, Tensor other) -> Tensor ");
  m.def("my_test_op_templated(Tensor self, Tensor other) -> Tensor ");
  m.def("my_test_view_op2_baseline(Tensor(a) self) -> Tensor(a) ");
  m.def("my_test_view_op2_boxed(Tensor(a) self) -> Tensor(a) ");
  m.def("my_test_view_op2_templated(Tensor(a) self) -> Tensor(a) ");
  m.def("my_test_view_op2_codegen(Tensor(a) self) -> Tensor(a) ");
  m.def("my_test_view_op2_fallthrough(Tensor(a) self) -> Tensor(a) ");
  m.def("my_test_inplace_op_baseline(Tensor(a!) self, Tensor other) -> Tensor(a!) ");
  m.def("my_test_inplace_op_boxed(Tensor(a!) self, Tensor other) -> Tensor(a!) ");
  m.def("my_test_inplace_op_templated(Tensor(a!) self, Tensor other) -> Tensor(a!) ");
  m.def("my_test_inplace_op_codegen(Tensor(a!) self, Tensor other) -> Tensor(a!) ");
  m.def("my_test_inplace_op_fallthrough(Tensor(a!) self, Tensor other) -> Tensor(a!) ");
}

TORCH_LIBRARY_IMPL(my_ops, CPU, m) {
  m.impl("my_test_op_baseline", TORCH_FN(my_test_op));
  m.impl("my_test_op_boxed", TORCH_FN(my_test_op));
  m.impl("my_test_op_templated", TORCH_FN(my_test_op));
  m.impl("my_test_view_op2_baseline", TORCH_FN(my_test_view_op2));
  m.impl("my_test_view_op2_boxed", TORCH_FN(my_test_view_op2));
  m.impl("my_test_view_op2_templated", TORCH_FN(my_test_view_op2));
  m.impl("my_test_view_op2_codegen", TORCH_FN(my_test_view_op2));
  m.impl("my_test_view_op2_fallthrough", TORCH_FN(my_test_view_op2));
  m.impl("my_test_inplace_op_baseline", TORCH_FN(my_test_inplace_op));
  m.impl("my_test_inplace_op_boxed", TORCH_FN(my_test_inplace_op));
  m.impl("my_test_inplace_op_templated", TORCH_FN(my_test_inplace_op));
  m.impl("my_test_inplace_op_codegen", TORCH_FN(my_test_inplace_op));
  m.impl("my_test_inplace_op_fallthrough", TORCH_FN(my_test_inplace_op_for_fallthrough));
}

/**
 * Register Autograd and ADInplaceOrView kernels
 * We always register boxed Autograd kernel, but for...
 *  - boxed => boxed kernel for ADInplaceOrView
 *  - templated => templated kernel for ADInplaceOrView
 *  - codegen => copy-pasted codegened kernel for ADInplaceOrview
 *  - fallthrough => don't register ADInplaceOrView
 */
REGISTER_AUTOGRAD_NOT_IMPLEMENTED_FALLBACK(my_ops, "my_test_op_boxed");
REGISTER_AUTOGRAD_NOT_IMPLEMENTED_FALLBACK2(my_ops, "my_test_op_templated", "my_ops::my_test_op_templated", "my_test_op_templated(Tensor self, Tensor other) -> Tensor ", my_test_op);
REGISTER_AUTOGRAD_NOT_IMPLEMENTED_FALLBACK(my_ops, "my_test_view_op2_boxed");
REGISTER_AUTOGRAD_NOT_IMPLEMENTED_FALLBACK2(my_ops, "my_test_view_op2_templated", "my_ops::my_test_view_op2_templated", "my_test_view_op2_templated(Tensor(a) self) -> Tensor(a)", my_test_view_op2);
REGISTER_AUTOGRAD_NOT_IMPLEMENTED_FALLBACK(my_ops, "my_test_inplace_op_boxed");
REGISTER_AUTOGRAD_NOT_IMPLEMENTED_FALLBACK2(my_ops, "my_test_inplace_op_templated", "my_ops::my_test_inplace_op_templated", "my_test_inplace_op_templated(Tensor(a!) self, Tensor other) -> Tensor(a!)", my_test_inplace_op);

TORCH_LIBRARY_IMPL(my_ops, Autograd, m) {
  m.impl("my_test_view_op2_codegen", autogradNotImplementedFallback());
}
// copy-pasted codegened InplaceOrView kernel, but instead of redispatching, call directly into the function
// this could make a difference?
TORCH_LIBRARY_IMPL(my_ops, ADInplaceOrView, m) {
  m.impl("my_test_view_op2_codegen", TORCH_FN(aliasInplaceOrViewCodegened));
}

// Only register Autograd kernel for this one
TORCH_LIBRARY_IMPL(my_ops, Autograd, m) {
  m.impl("my_test_view_op2_fallthrough", autogradNotImplementedFallback());
}

TORCH_LIBRARY_IMPL(my_ops, Autograd, m) {
  m.impl("my_test_inplace_op_codegen", autogradNotImplementedFallback());
}
// copy-pasted codegened InplaceOrView kernel, but instead of redispatching, call directly into the function
// this could make a difference?
TORCH_LIBRARY_IMPL(my_ops, ADInplaceOrView, m) {
  m.impl("my_test_inplace_op_codegen", TORCH_FN(addInplaceCodegened));
}

// Only register Autograd kernel for this one
TORCH_LIBRARY_IMPL(my_ops, Autograd, m) {
  m.impl("my_test_inplace_op_fallthrough", autogradNotImplementedFallback());
}

// TORCH_LIBRARY_IMPL(my_ops, ADInplaceOrView, m) {
//   m.impl("my_test_inplace_op_fallthrough", TORCH_FN(my_test_inplace_op_fallthrough));
// }

static void baseline(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_baseline(a, b);
  }
}
BENCHMARK(baseline);

static void boxed(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_boxed(a, b);
  }
}
BENCHMARK(boxed);

static void templated(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_templated(a, b);
  }
}
BENCHMARK(templated);

static void baseline_view(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  for (auto _ : state) {
    dispatch_to_baseline_view(a);
  }
}
BENCHMARK(baseline_view);

static void boxed_view(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  for (auto _ : state) {
    dispatch_to_boxed_view(a);
  }
}
BENCHMARK(boxed_view);

static void templated_view(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  for (auto _ : state) {
    dispatch_to_templated_view(a);
  }
}
BENCHMARK(templated_view);

static void codegen_view(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  for (auto _ : state) {
    dispatch_to_codegen_view(a);
  }
}
BENCHMARK(codegen_view);

static void fallthrough_view(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true);
  for (auto _ : state) {
    dispatch_to_fallthrough_view(a);
  }
}
BENCHMARK(fallthrough_view);

static void baseline_inplace(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true).clone();
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_baseline_inplace(a, b);
  }
}
BENCHMARK(baseline_inplace);

static void boxed_inplace(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true).clone();
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_boxed_inplace(a, b);
  }
}
BENCHMARK(boxed_inplace);

static void templated_inplace(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true).clone();
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_templated_inplace(a, b);
  }
}
BENCHMARK(templated_inplace);

static void codegen_inplace(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true).clone();
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_codegen_inplace(a, b);
  }
}
BENCHMARK(codegen_inplace);

static void fallthrough_inplace(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true).clone();
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_fallthrough_inplace(a, b);
  }
}
BENCHMARK(fallthrough_inplace);

static void fallthrough_inplace_a_second_time(benchmark::State& state) {
  torch::Tensor a = torch::ones({1, 2, 3}).set_requires_grad(true).clone();
  torch::Tensor b = torch::ones({1, 2, 3}).set_requires_grad(false);
  for (auto _ : state) {
    dispatch_to_fallthrough_inplace(a, b);
  }
}
BENCHMARK(fallthrough_inplace_a_second_time);


BENCHMARK_MAIN();

```
EDIT: updated benchmark and code to include the kernel copy-pasted from codegened code

</details>